Update README.md
Changed the description to reflect the changes in framework: instead of loading all data, we only work on an input position, loading only the archival MasterSource that is of interest. This lead to changing the names of the Python scripts to reflect this change, and removing the StudyMasterSource script, which does not work anymore in this new framework.