Add files via upload
New Python scripts that reflect the change in framework. Instead of loading all data, we work focused on a single input position and flux state. The selection of a sub-sample of the MasterSource catalog relevant for a given observation is done only once, the first time the ObsID is met.